# swagger.yaml
openapi: 3.0.0
info:
  title: Kawan Nusa API
  description: Official API Documentation for Kawan Nusa Backend Service. Includes all endpoints for Authentication, Profile, Points, Customers, Rewards, Redemptions, Statistics, Education, Catalog, and more.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: /api
    description: Relative API server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. Expires in 15 minutes."
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: "Internal API key for server-to-server communication."
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        statusCode: { type: integer, example: 200 }
        message: { type: string, example: "Operation successful" }
...
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          { type: string, format: email, example: "budi.santoso@nusanet.id" }
...
    AuthUser:
      type: object
      properties:
        id: { type: integer, example: 1 }
        name: { type: string, example: "Budi Santoso" }
        photo:
          {
            type: string,
            nullable: true,
            example: "https://cdn.example.com/photo.jpg",
          }
        email:
          { type: string, format: email, nullable: true, example: "budi.santoso@nusanet.id" }
        phone: { type: string, nullable: true, example: "081234567890" }
        isActive: { type: boolean, example: true }
        role:
          type: string
          enum: ["user", "admin"]
          example: "user"
...
  responses:
...
    ValidationErrorResponse:
      description: Validation Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
paths:
...
  /auth/register:
    post:
      summary: Register new partner account
      description: "Register a new partner account with personal info, bank details, and document uploads (KTP & bank book). Account will be set as inactive (isActive=false) until approved by admin."
      tags: [Auth]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthUser" }
              example:
                success: true
                statusCode: 201
                message: "User registered successfully"
                data:
                  id: 1
                  name: "Budi Santoso"
                  photo: null
                  email: "budi.santoso@nusanet.id"
                  phone: "081234567890"
                  isActive: false
                  role: "user"
        "400":
          description: Email or phone already in use
          content:
            application/json:
              example:
                success: false
                statusCode: 400
                message: "Email already in use"
        "422":
          $ref: "#/components/responses/ValidationErrorResponse"
  /auth/verify-email:
    get:
      summary: Verify email and auto-login
      description: "Verify the user's email using the token sent via email. Sets isVerified=true, isActive=true, and returns auth tokens for auto-login."
      tags: [Auth]
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
          description: "Email verification token"
      responses:
        "200":
          description: Email verified and logged in successfully
          content:
            application/json:
              example:
                success: true
                statusCode: 200
                message: "Email verified successfully"
                data:
                  user:
                    id: 1
                    name: "John Doe"
                    photo: null
                    email: "john@example.com"
                    phone: "081234567890"
                    isActive: true
                    isBoarding: false
                    role: "user"
                  accessToken: "eyJhbGciOiJIUzI1NiIs..."
                  refreshToken: "eyJhbGciOiJIUzI1NiIs..."
        "400":
          description: Invalid/expired token or already verified
          content:
            application/json:
              example:
                success: false
                statusCode: 400
                message: "Invalid or expired verification token"
  /auth/resend-verification:
    post:
      summary: Resend email verification
      description: "Resend the verification email to an unverified user."
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, example: "budi.santoso@nusanet.id" }
      responses:
        "200":
          description: Verification email sent successfully
          content:
            application/json:
              example:
                success: true
                statusCode: 200
                message: "Verification email sent successfully"
                data: null
        "400":
          description: Email not found or already verified
          content:
            application/json:
              example:
                success: false
                statusCode: 400
                message: "Email already verified"
        "422":
          $ref: "#/components/responses/ValidationErrorResponse"
  /auth/check-email-status:
    get:
      summary: Check email verification status
      description: "Lightweight check to see if an email needs verification. Does not send any email."
      tags: [Auth]
      parameters:
        - name: email
          in: query
          required: true
          schema: { type: string, format: email }
          description: "Email address to check"
      responses:
        "200":
          description: Email needs verification
          content:
            application/json:
              example:
                success: true
                statusCode: 200
                message: "Success"
                data:
                  needsVerification: true
        "400":
          description: Email not found or already verified
          content:
            application/json:
              example:
                success: false
                statusCode: 400
                message: "Email already verified"
  /auth/login:
    post:
      summary: Login to partner account
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
            example:
              identifier: "budi.santoso@nusanet.id"
              password: "secret123"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
              example:
                success: true
                statusCode: 200
                message: "Logged in successfully"
                data:
                  user:
                    id: 1
                    name: "Budi Santoso"
                    photo: null
                    email: "budi.santoso@nusanet.id"
                    phone: "081234567890"
                    isActive: true
                    role: "user"
                  accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "401":
          description: Invalid credentials
          content:
            application/json:
              example:
                success: false
                statusCode: 401
                message: "Invalid credentials"
        "422":
          $ref: "#/components/responses/ValidationErrorResponse"
  /auth/google:
    post:
      summary: Login with Google OAuth (user)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, example: "4/0AX4XfWh..." }
            example:
              code: "4/0AX4XfWh..."
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
              example:
                success: true
                statusCode: 200
                message: "Logged in successfully"
                data:
                  user:
                    id: 1
                    name: "Budi Santoso"
                    photo: null
                    email: "budi.santoso@nusanet.id"
                    phone: "081234567890"
                    isActive: true
                    role: "user"
                  accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "400":
          description: User not registered
          content:
            application/json:
              example:
                success: false
                statusCode: 400
                message: "User not registered"
  /auth/admin/google:
    post:
      summary: Login with Google OAuth (admin/employee)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, example: "4/0AX4XfWh..." }
            example:
              code: "4/0AX4XfWh..."
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
              example:
                success: true
                statusCode: 200
                message: "Logged in successfully"
                data:
                  user:
                    id: 10
                    name: "Admin User"
                    photo: null
                    email: "admin@nusanet.id"
                    phone: "081234567890"
                    isActive: true
                    role: "admin"
                  accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "400":
          description: Employee not registered
          content:
            application/json:
              example:
                success: false
                statusCode: 400
                message: "Employee not registered"
  /auth/forgot-password:
    post:
      summary: Request password reset link via email
      description: |
        Generates a new reset token stored in a separate `password_reset_tokens` table.
        Multiple tokens can exist per user (each forgot-password request creates a new one).
        Tokens expire after 10 hours.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForgotPasswordRequest" }
            example:
              email: "budi.santoso@nusanet.id"
      responses:
        "200":
          description: Reset instructions sent to email
          content:
            application/json:
              example:
                success: true
                statusCode: 200
                message: "Password reset instructions have been sent to your email"
                data: null
        "400":
          description: Email not found
          content:
            application/json:
              example:
                success: false
                statusCode: 400
                message: "Email not found"
        "422":
          $ref: "#/components/responses/ValidationErrorResponse"
  /auth/validate-reset-token:
    get:
      summary: Validate a password reset token
      description: |
        Validates the token from the `password_reset_tokens` table.
        Only the token is required — the user is resolved from the token's relation.
      tags: [Auth]
      parameters:
        - name: token
          in: query
          required: true
          description: Reset token received via email
          schema: { type: string }
          example: "a3f8c2d1e9b4..."
      responses:
        "200":
          description: Token is valid
          content:
            application/json:
              example:
                success: true
                statusCode: 200
                message: "Token is valid"
                data: null
        "400":
          description: Invalid or expired token
          content:
            application/json:
              example:
                success: false
                statusCode: 400
                message: "Invalid or expired reset token"
  /auth/reset-password:
    post:
      summary: Reset password using token from email
      description: |
        Validates the token from the `password_reset_tokens` table, updates the user's password,
        then deletes **all** reset tokens for that user.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordRequest" }
            example:
              token: "a3f8c2d1e9b4..."
              newPassword: "newSecret123"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              example:
                success: true
                statusCode: 200
                message: "Password has been successfully reset"
                data: null
        "400":
          description: Invalid or expired token
          content:
            application/json:
              example:
                success: false
                statusCode: 400
                message: "Invalid or expired reset token"
        "422":
          $ref: "#/components/responses/ValidationErrorResponse"
  /auth/refresh:
    post:
      summary: Refresh access token using refresh token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshTokenRequest" }
            example:
              refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
              example:
                success: true
                statusCode: 200
                message: "Token refreshed successfully"
                data:
                  user:
                    id: 1
                    name: "Budi Santoso"
                    photo: null
                    email: "budi.santoso@nusanet.id"
                    phone: "081234567890"
                    isActive: true
                    role: "user"
                  accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.new..."
                  refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.new..."
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              example:
                success: false
                statusCode: 401
                message: "Invalid or expired refresh token"
  /auth/me:
    get:
      summary: Get authenticated user details
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: User profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthUser" }
              example:
                success: true
                statusCode: 200
                message: "User profile retrieved successfully"
                data:
                  id: 1
                  name: "Budi Santoso"
                  photo: null
                  email: "budi.santoso@nusanet.id"
                  phone: "081234567890"
                  isActive: true
                  role: "user"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /auth/logout:
    post:
      summary: Logout (client should discard tokens)
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              example:
                success: true
                statusCode: 200
                message: "Logged out successfully"
                data: null
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /auth/otp/send:
    post:
      summary: Send OTP code for login
      description: "Sends a 6-digit OTP code to the user's email or phone. If identifier contains '@', sends via email. Otherwise, logs to console (phone not yet implemented)."
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              properties:
                identifier:
                  type: string
                  description: "Email address or phone number"
                  example: "user@example.com"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              example:
                success: true
                statusCode: 200
                message: "OTP sent successfully"
                data:
                  type: "email"
        "400":
          description: User not found, not verified, or inactive
          content:
            application/json:
              example:
                success: false
                statusCode: 400
                message: "User not found"
...